Basic Code Done
I haven't committed yet, but I'm done with the basic HTML, CSS, and JS files, which are all part of this commit. At the moment, the program is not very visual, but its calculations are correct, and account for 4 or 5 bands. I'll add a six band update later, along with another box to reverse determine colors.
Pushing mainly because I'm gonna' start adding some visuals!