Color Display
Added the code to display the colors! Updated the color objects with their respective names. Gonna' have to change up the colors, though, since a) I don't like most of them, and b) grey is too similar to silver, while gold is too similar to yellow.

Regardless, it's working pretty well so far!